+++ app/main.cpp
@@ -1,4 +1,12 @@
 #include "term.h"
+#include <iostream>
+#include <clocale>
+#include <math.h>
+#include <string>
+#include <fstream>
+#include <cstdlib>
+#include <Windows.h>
+
 using namespace std;
 
 int main() {
